Split code of import js
