+++ tools/importer/transformers/accordion.js
@@ -1,0 +1,23 @@
+
+const createAccordion = (main, document) => {
+    const accordion = main.querySelector('accordion');
+    const cells = [['Accordion']];
+    if (accordion) {
+      const accordionHeader = document.createElement('div');
+      accordionHeader.textContent = accordion.getAttribute('accordionheader');
+      const accordionLists = JSON.parse(decodeHtmlEntities(accordion.getAttribute('accordionlist')));
+      const definitionlists = accordionLists.map((list) => {
+        const pEl = document.createElement('p');
+        pEl.innerHTML = list.description;
+        const divEl = document.createElement('div');
+        divEl.innerHTML = list.title;
+        divEl.append(pEl);
+        return [divEl];
+      });
+      if (accordionHeader.textContent) cells.push([accordionHeader]);
+      cells.push(...definitionlists);
+      const block = WebImporter.DOMUtils.createTable(cells, document);
+      main.append(block);
+    }
+  };
+  export default createAccordion;
