+++ tools/importer/transformers/breadcrumb.js
@@ -1,0 +1,30 @@
+const createBreadcrumb = (main, document) => {
+    const breadcrumb = main.querySelector('div.breadcrumb');
+    if (breadcrumb) {
+      const breadcrumbEl = breadcrumb.querySelector('breadcrumb');
+      if (breadcrumbEl) {
+        const cells = [];
+        // eslint-disable-next-line no-undef
+        const list = JSON.parse(decodeHtmlEntities(breadcrumbEl.getAttribute('breadcrumbdetailslist')));
+        cells.push(['Breadcrumb']);
+        const ul = document.createElement('ul');
+        list.forEach((item) => {
+          if (!item.url?.includes('/content/experience-fragments')) {
+            const li = document.createElement('li');
+            const anc = document.createElement('a');
+            anc.href = item.url;
+            anc.textContent = item.title;
+            li.append(anc);
+            ul.append(li);
+          }
+        });
+        cells.push([ul]);
+        if (cells.length > 0 && ul.firstElementChild) {
+          const block = WebImporter.DOMUtils.createTable(cells, document);
+          const firstChild = main.firstElementChild?.firstChild;
+          main.firstElementChild.insertBefore(block, firstChild);
+        }
+      }
+    }
+  };
+export default createBreadcrumb;
